Use common code for executable, move Cnd to it
