+++ src/executable/btsieve.rs
@@ -78,9 +78,7 @@ impl Default for LogLevels {
 
 impl Default for HttpApi {
     fn default() -> HttpApi {
-        let port_bind = port_check::free_local_port()
-            .expect("Could not find a free port")
-            .into();
+        let port_bind = port_check::free_local_port().expect("Could not find a free port");
         HttpApi {
             address_bind: IpAddr::V4(Ipv4Addr::UNSPECIFIED),
             port_bind,
