+++ src/lib.rs
@@ -1,6 +1,5 @@
 pub mod bitcoin;
-pub mod btsieve;
-pub mod cnd;
 pub mod create_comit_app;
 pub mod ethereum;
+pub mod executable;
 pub mod start_env;
